+++ src/tensorwaves/estimator.py
@@ -6,6 +6,7 @@ from typing import Callable, Dict, Mapping, Optional, Union
 
 import numpy as np
 
+from tensorwaves.backend import find_function
 from tensorwaves.interface import (
     DataSample,
     Estimator,
@@ -14,7 +15,6 @@ from tensorwaves.interface import (
     ParameterValue,
 )
 from tensorwaves.model import LambdifiedFunction
-from tensorwaves.model.backend import find_function
 
 
 def gradient_creator(
